+++ src/Framework/Bootstrap/AbstractSetupGacela.php
@@ -33,6 +33,8 @@ abstract class AbstractSetupGacela implements SetupGacelaInterface
 
     public const contextualBindings = 'contextualBindings';
 
+    public const lazyServices = 'lazyServices';
+
     public const plugins = 'plugins';
 
     public const gacelaConfigsToExtend = 'gacelaConfigsToExtend';
@@ -62,6 +64,8 @@ abstract class AbstractSetupGacela implements SetupGacelaInterface
     protected const DEFAULT_ALIASES = [];
 
     protected const DEFAULT_CONTEXTUAL_BINDINGS = [];
+
+    protected const DEFAULT_LAZY_SERVICES = [];
 
     protected const DEFAULT_GACELA_CONFIGS_TO_EXTEND = [];